Completata la sezione per le Pizzerie.

+++ app/database/query.sql
@@ -2,7 +2,7 @@
  * Creazione tabella Cliente
  */
 CREATE TABLE IF NOT EXISTS Cliente(
-	IDCliente INTEGER(11) PRIMARY KEY,
+	IDCliente INTEGER PRIMARY KEY,
 	Nome_Cl VARCHAR(50) NOT NULL,
 	Cognome_Cl VARCHAR(30) NOT NULL
 );
@@ -11,7 +11,7 @@ CREATE TABLE IF NOT EXISTS Cliente(
  * Creazione tabella Prodotto.
  */
 CREATE TABLE IF NOT EXISTS Prodotto(
-	IDProdotto INTEGER(11) PRIMARY KEY,
+	IDProdotto INTEGER PRIMARY KEY,
 	Nome_Pr VARCHAR(30) NOT NULL
 );
 
@@ -19,8 +19,8 @@ CREATE TABLE IF NOT EXISTS Prodotto(
  * Creazione tabella Ordine (associazione tra Cliente e Prodotto).
  */
 CREATE TABLE IF NOT EXISTS Ordine(
-	IDCliente INTEGER(11),
-	IDProdotto INTEGER(11),
+	IDCliente INTEGER,
+	IDProdotto INTEGER,
 	PRIMARY KEY(IDCliente, IDProdotto)
 );
 
@@ -29,7 +29,7 @@ CREATE TABLE IF NOT EXISTS Ordine(
  * Creazione tabella Ingrediente.
  */
 CREATE TABLE IF NOT EXISTS Ingrediente(
-	IDIngrediente INTEGER(11) PRIMARY KEY,
+	IDIngrediente INTEGER PRIMARY KEY,
 	Nome_Ing VARCHAR(30) NOT NULL
 );
 
@@ -37,8 +37,8 @@ CREATE TABLE IF NOT EXISTS Ingrediente(
  * Creazione tabella CompostoDa (associazione tra Prodotto e Ingrediente).
  */
 CREATE TABLE IF NOT EXISTS CompostoDa(
-	IDProdotto INTEGER(11),
-	IDIngrendiente INTEGER(11),
+	IDProdotto INTEGER,
+	IDIngrendiente INTEGER,
 	PRIMARY KEY(IDProdotto, IDIngrendiente)
 );
 
@@ -46,7 +46,7 @@ CREATE TABLE IF NOT EXISTS CompostoDa(
  * Creazione tabella Pizzeria.
  */
 CREATE TABLE IF NOT EXISTS Pizzeria(
-	IDPizzeria INTEGER(11) PRIMARY KEY,
+	IDPizzeria INTEGER PRIMARY KEY,
 	Nome_Pizz VARCHAR(40) NOT NULL,
 	Numero_Tel VARCHAR(30) NOT NULL,
 	CAP VARCHAR(5) NOT NULL,
@@ -60,8 +60,8 @@ CREATE TABLE IF NOT EXISTS Pizzeria(
  * Creazione tabella Appartiene (associazione tra Prodotto e Pizzeria).
  */
 CREATE TABLE IF NOT EXISTS Appartiene(
-	IDProdotto INTEGER(11),
-	IDPizzeria INTEGER(11),
+	IDProdotto INTEGER,
+	IDPizzeria INTEGER,
 	PRIMARY KEY(IDProdotto, IDPizzeria)
 );
